Extract CLI commands, prompts, context, and tool barrel; slim index.ts to pure wiring
Phase 3:
- Extract migrate CLI to src/cli/migrate.ts
- Extract import-claude-memory CLI to src/cli/import-claude-memory.ts
- Extract prompt registrations to src/prompts.ts

Phase 4:
- Extract createServerContext and readPackageVersion to src/context.ts
- Create src/tools/index.ts barrel for all tool registrations
- Extend src/startup.ts with startServer()
- Slim src/index.ts from ~589 to 41 lines
