+++ src/startup.ts
@@ -1,5 +1,7 @@
 import { readVaultSchemaVersion } from "./config.js";
 import type { ServerContext } from "./server-context.js";
+import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
+import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
 export async function warnAboutPendingMigrationsOnStartup(ctx: ServerContext): Promise<void> {
   let totalPending = 0;
@@ -29,4 +31,20 @@ export async function warnAboutPendingMigrationsOnStartup(ctx: ServerContext): P
   for (const detail of details) {
     console.error(`[mnemonic]   ${detail}`);
   }
+}
+
+export async function startServer(server: McpServer, ctx: ServerContext): Promise<void> {
+  await warnAboutPendingMigrationsOnStartup(ctx);
+  const transport = new StdioServerTransport();
+
+  async function shutdown() {
+    await server.close();
+    process.exit(0);
+  }
+  process.on("SIGINT", shutdown);
+  process.on("SIGTERM", shutdown);
+  transport.onclose = async () => { await server.close(); };
+
+  await server.connect(transport);
+  console.error(`[mnemonic] Started. Main vault: ${ctx.vaultPath}`);
 }
